src/requirements.txt
python-binance==1.0.16
requests==2.27.1
retry==0.9.2
prettytable==3.6.0
